fix: 🐛 non-critical circular deps
