+++ src/components/Collapse/types/CollapseIconFactoryType.ts
@@ -1,0 +1,9 @@
+import { ReactNode } from 'react';
+
+export type CollapseIconFactoryType = (
+    iconProps: CollapseIconFactoryTypeArgs
+) => ReactNode;
+
+export type CollapseIconFactoryTypeArgs = {
+    isActive: boolean;
+};
