+++ src/components/Collapse/types/index.ts
@@ -1,23 +1,16 @@
-import { PropsWithChildren, ReactNode } from 'react';
+import { PropsWithChildren } from 'react';
 import { CollapseProviderProps } from '../context/types';
 import { CommonHTMLProps } from '../../../types';
+import { CollapseIconFactoryType } from './CollapseIconFactoryType';
+
+export * from './CollapseIconFactoryType';
 
 export type StyledCollapseProps = {};
-
-export type ControllerCollapseProps = CollapseProviderProps;
 
 export type CollapseProps = PropsWithChildren<
     CommonHTMLProps<HTMLDivElement> &
         StyledCollapseProps &
-        ControllerCollapseProps & {
+        CollapseProviderProps & {
             expandIcon?: CollapseIconFactoryType;
         }
 >;
-
-export type CollapseIconFactoryType = (
-    iconProps: CollapseIconFactoryTypeArgs
-) => ReactNode;
-
-export type CollapseIconFactoryTypeArgs = {
-    isActive: boolean;
-};
