+++ src/components/Typography/Text/types/TextElement.tsx
@@ -1,0 +1,1 @@
+export type TextElement = 'span' | 'strong' | 'code' | 'del' | 'mark';
