+++ src/components/Typography/Text/types/TextProps.tsx
@@ -1,0 +1,7 @@
+import { CommonHTMLProps, TypeWithDocs } from '../../../../types';
+import { TextElement } from './TextElement';
+
+export type TextProps = {
+    element?: TextElement;
+} & CommonHTMLProps &
+    TypeWithDocs;
