+++ src/components/Typography/Text/types/index.ts
@@ -1,0 +1,2 @@
+export * from './TextElement';
+export * from './TextProps';
